Change implementation to work with Either instead of Maybe

+++ src/Lib/Types.hs
@@ -12,7 +12,7 @@ data NOAAResult = NOAAResult
   , mindate :: T.Text
   , maxdate :: T.Text
   , name :: T.Text
-  , datacoverage :: Int
+  , datacoverage :: Double
   , resultId :: T.Text
   } deriving Show
